Updated Equipment Insatnce Resolvers

+++ server/src/resolvers/equipmentInstanceResolver.ts
@@ -1,4 +1,5 @@
 import * as EquipmentRepo from "../repositories/Equipment/EquipmentRepository.js";
+import * as RoomRepo from "../repositories/Rooms/RoomRepository.js"
 import { Privilege } from "../schemas/usersSchema.js";
 import { ApolloContext } from "../context.js";
 import { getAccessCheckByID, getAccessChecks, getAccessChecksByApproved, setAccessCheckApproval } from "../repositories/Equipment/AccessChecksRepository.js";
@@ -32,8 +33,8 @@ const EquipmentInstanceResolver = {
     equipmentInstances: async (
       _parent: any,
       args: { equipmentID: number },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async () => {
+      { isStaff }: ApolloContext) =>
+      isStaff(async () => {
         return await getInstancesByEquipment(args.equipmentID)
       }),
   },
@@ -49,8 +50,8 @@ const EquipmentInstanceResolver = {
     createEquipmentinstance: async (
       _parent: any,
       args: { equipmentID: number, name: string },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async (user) => {
+      { isManager }: ApolloContext) =>
+      isManager(async (user) => {
         const equipment = await EquipmentRepo.getEquipmentByID(args.equipmentID);
         if (!equipment) throw new GraphQLError("Equipment does not exist");
         await createLog(`{user} created instance "${args.name}" on {equipment}`, "admin", { id: user.id, label: getUsersFullName(user) }, { id: equipment.id, label: equipment.name });
@@ -67,12 +68,16 @@ const EquipmentInstanceResolver = {
     setInstanceStatus: async (
       _parent: any,
       args: { id: number, status: string },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async (user) => {
+      { isStaff }: ApolloContext) =>
+      isStaff(async (user) => {
         const orig = await getInstanceByID(args.id);
         if (!orig) throw new GraphQLError("Instance does not exist");
         const equipment = await EquipmentRepo.getEquipmentByID(orig.equipmentID);
         if (!equipment) throw new GraphQLError("Equipment does not exist");
+        const room = await RoomRepo.getRoomByID(equipment.roomID);
+        if (!user.staff.includes(room?.zoneID ?? -1) && !user.manager.includes(room?.zoneID ?? -1) && !user.admin) {
+          throw new GraphQLError(`Not Privileged for Makerspace ${room?.zoneID}`);
+        }
         await createLog(`{user} changed instance "${orig.name}" status to "${args.status}" on {equipment}`, "admin", { id: user.id, label: getUsersFullName(user) }, { id: equipment.id, label: equipment.name });
         return await setInstanceStatus(args.id, args.status)
       }),
@@ -87,12 +92,16 @@ const EquipmentInstanceResolver = {
     setInstanceName: async (
       _parent: any,
       args: { id: number, name: string },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async (user) => {
+      { isManager }: ApolloContext) =>
+      isManager(async (user) => {
         const orig = await getInstanceByID(args.id);
         if (!orig) throw new GraphQLError("Instance does not exist");
         const equipment = await EquipmentRepo.getEquipmentByID(orig.equipmentID);
         if (!equipment) throw new GraphQLError("Equipment does not exist");
+        const room = await RoomRepo.getRoomByID(equipment.roomID);
+        if (!user.manager.includes(room?.zoneID ?? -1) && !user.admin) {
+          throw new GraphQLError(`Not Privileged for Makerspace ${room?.zoneID}`);
+        }
         await createLog(`{user} changed instance "${orig.name}" name to "${args.name}" on {equipment}`, "admin", { id: user.id, label: getUsersFullName(user) }, { id: equipment.id, label: equipment.name });
         return await setInstanceName(args.id, args.name)
       }),
@@ -106,12 +115,16 @@ const EquipmentInstanceResolver = {
     deleteInstance: async (
       _parent: any,
       args: { id: number },
-      { ifAllowed }: ApolloContext) =>
-      ifAllowed([Privilege.MENTOR, Privilege.STAFF], async (user) => {
+      { isManager }: ApolloContext) =>
+      isManager(async (user) => {
         const orig = await getInstanceByID(args.id);
         if (!orig) throw new GraphQLError("Instance does not exist");
         const equipment = await EquipmentRepo.getEquipmentByID(orig.equipmentID);
         if (!equipment) throw new GraphQLError("Equipment does not exist");
+        const room = await RoomRepo.getRoomByID(equipment.roomID);
+        if (!user.manager.includes(room?.zoneID ?? -1) && !user.admin) {
+          throw new GraphQLError(`Not Privileged for Makerspace ${room?.zoneID}`);
+        }
         await createLog(`{user} deleted instance "${orig.name}" on {equipment}`, "admin", { id: user.id, label: getUsersFullName(user) }, { id: equipment.id, label: equipment.name });
         return await deleteInstance(args.id)
       }),
